requirements.txt
fastapi[standard]~=0.115.8
fastapi-jwt
pydantic-settings~=2.7.1
sqlalchemy~=2.0.38
alembic~=1.14.1
asyncpg~=0.30.0
starlette~=0.45.3
pydantic~=2.10.6
python-dotenv~=1.0.1
passlib
bcrypt
authlib
orjson